fix: 24 - UNBOX_NULLABLE - return (t1 - t2); - unboxing of a @Nullable value - t2

+++ src/main/java/com/ibm/wala/util/graph/traverse/DFS.java
@@ -132,7 +132,7 @@ public class DFS {
       Integer t1 = order.get(o1);
       Integer t2 = order.get(o2);
       // throws an exception if either node has not been ordered
-      return (t1 - t2);
+      return (t1.intValue() - t2.intValue());
     }
   }
